+++ packages/office-addin-usage-data/src/defaults.ts
@@ -1,5 +1,5 @@
 import * as os from "os";
 import * as path from "path";
 
-export const usageDataJsonFilePath: string = path.join(os.homedir(), "/officeAddinUsageData.json");
-export const groupName = "Office-Addin-Usage-Data";
+export const usageDataJsonFilePath: string = path.join(os.homedir(), "/office-addin-usage-data.json");
+export const groupName = "office-addin-usage-data";
